Revert "fix: Check if any nodes relate to proposal.projectType (#846)"
This reverts commit 7a8dddb76212ff02a6b59968dcf575b31583e6dc.
